backend/pyproject.toml
[project]
name = "whatsapp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name ="marktecs.com", company = "marktecs.com"},
    { name = "Muhammad Fasih", email = "programmer.fasih.18@gmail.com", github = "HafizFasih" },
    { name = "Anas Ahmed Shaikh", email = "anasahmedshaikh.work@gmail.com", github = "anasahmed07" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.6",
    "apscheduler>=3.11.0",
    "asyncpg>=0.29.0",
    "bcrypt>=4.3.0",
    "docx2txt>=0.9",
    "fastapi>=0.115.12",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "intuit-oauth>=1.2.6",
    "kafka>=1.3.5",
    "oauth2client>=4.1.3",
    "openai>=1.99.3",
    "openai-agents>=0.0.16",
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.13.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "pywa>=3.0.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "supabase>=2.17.0",
    "typer>=0.19.2",
    "tzdata>=2025.2",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]

[project.scripts]
whatsapp-agent = "whatsapp_agent.main:app"
run-server = "uvicorn:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
